requirements.txt
chardet
patool
